+++ src/Fretboarder/Guitar/Note.hs
@@ -66,7 +66,7 @@ fromINote :: INote -> Note
 fromINote i = Note t o a
   where
     i'     = i + 9
-    o      = floor $ realToFrac i' / 12.0
+    o      = floor $ fromIntegral i' / 12.0
     (t, a) = case i' `mod` 12 of
       0  -> (C, Natural)
       1  -> (C, Sharp)
